+++ app/src/main/java/com/rolandvitezhu/todocloud/network/api/category/dto/UpdateCategoryRequest.kt
@@ -15,7 +15,7 @@ data class UpdateCategoryRequest (
         @SerializedName("deleted")
         var deleted: Boolean?,
         @SerializedName("position")
-        var position: Int?
+        var position: Double?
 ) : Parcelable {
     constructor() : this(
             null,
