+++ app/src/main/java/com/rolandvitezhu/todocloud/network/api/list/dto/InsertListRequest.kt
@@ -17,7 +17,7 @@ data class InsertListRequest (
         @SerializedName("deleted")
         var deleted: Boolean?,
         @SerializedName("position")
-        var position: Int?
+        var position: Double?
 ) : Parcelable {
     constructor() : this(
             null,
