+++ app/src/main/java/com/rolandvitezhu/todocloud/network/api/todo/dto/UpdateTodoRequest.kt
@@ -27,7 +27,7 @@ data class UpdateTodoRequest (
         @SerializedName("deleted")
         var deleted: Boolean?,
         @SerializedName("position")
-        var position: Int?
+        var position: Double?
 ) : Parcelable {
     constructor() : this(
             null,
